third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-019-expected.txt


FAIL CSS Test: @viewport constrained - Auto width from height and non-auto zoom. assert_equals: expected 10000 but got 600
